Update to v0.87
- Added SWL mode: save and export SWL log via Web UI
SWL QSO will be logged when both stations provide signal report and 73/RR73/RRR (Gridsqure is optional)

- Added option to display log loopup in map view
- Added meter, monitor and control for Flexradio (NO TX yet)
- Added auto time sync (via Microsoft NTP)
- Added support for upcoming FW on Xiegu G90s

- Enhanced data query in backend Web UI
- Optimized runtime for log loopup
- Fixed distance mismatch in "Callsign and Gridsquare" on Web UI
- Improved performance to handle massive decoding

+++ ft8CN/app/src/main/java/com/bg7yoz/ft8cn/connector/BaseRigConnector.java
@@ -1,5 +1,10 @@
 package com.bg7yoz.ft8cn.connector;
-
+/**
+ * 用于连接电台的基础类，蓝牙、USB线、FLEX网络、ICOM网络都是继承于此
+ *
+ * @author BG7YOZ
+ * @date 2023-03-20
+ */
 
 import com.bg7yoz.ft8cn.rigs.OnConnectReceiveData;
 import com.bg7yoz.ft8cn.rigs.OnRigStateChanged;
@@ -69,7 +74,7 @@ public class BaseRigConnector {
         onConnectReceiveData=receiveData;
     }
 
-    public void sendWaveData(short[] data){
+    public void sendWaveData(float[] data){
         //留给网络方式发送音频流
     }
